requirements.txt
gunicorn
psycopg2-binary
Flask
python-dotenv==1.2.1